--- src/assets/css/footer-section.css
@@ -1,27 +1,0 @@
-.media-icon{
-    height: 24px ;
-}
-
-.media-icon:hover{
-    filter: contrast(50%);
-}
-
-.background-image{
-    background-image: url('/public/images/BottomBanner.svg');
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    gap: 80px;
-  }
-
-.logo-r-brown{
-    height: 100px ;
-    width: 100px;
-}
-.logo-dark-fit{
-    height: 48px;
-    width: 192px;
-}
