+++ src/assets/tailwind.output.css
@@ -29,4 +29,3 @@
 @import './css/body-heading.css';
 @import './css/chat-bubble.css';
 @import './css/testimonials.css';
-@import './css/footer-section.css';
