--- src/pages/links/faqsPage.js
@@ -1,191 +1,0 @@
-import React from "react";
-// import LogoRBrown from "../../../image/logoRBrown";
-import FooterSection from "../components/FooterSection";
-
-const FAQsPage = () => {
-  return (
-    <div class="mx-auto bg-dark-light">
-      <div class="flex justify-center max-w-[100px] sm:max-w-[1120px]">
-        <img
-          src="./images/logo_r_brown.svg"
-          alt="logo_brown"
-          className="logo-r-brown"
-        />
-      </div>
-      <hr />
-      <div class="flex justify-center">
-        <div class="max-w-4xl">
-          <h1 class="font-inter font-bold text-4xl text-brown-900 p-6">
-            FAQs About Pet Adoption
-          </h1>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            I’ve found a pet I’m interested in. Now what?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Congratulations! Your next step is to contact the organization
-            listed on the pet’s page. Click “learn more about me” to go to
-            Adopt-a-Pet.com for additional information on the shelter or rescue
-            and its adoption process.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            What is the difference between a shelter and a rescue group?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            The biggest difference is that most shelters keep their available
-            pets in a single facility, while most rescue groups keep theirs in
-            the homes of volunteers.{" "}
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Also, shelters usually have paid staff as well as volunteers, while
-            rescue groups are typically volunteer-run efforts.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            The pet I’m interested in is at a shelter; what will it be like when
-            I go to see her?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Shelters come in all varieties. Some are modern and beautiful, while
-            others are old-fashioned. Some are run by private charities (and
-            usually have terms like “SPCA” or “Humane Society” in their names)
-            and others are government facilities and may be what you’d call
-            “no-frills.”
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            After you’ve selected a pet who seems right for you, most private
-            shelters will let you visit with him or her in their own “dog
-            apartments” or “kitty condos,” or will invite you to spend time with
-            the pet in a special room or fenced outdoor area set aside for pets
-            and adopters to get acquainted.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            If you are interested in a dog and already have a dog at home, many
-            private shelters will require you to bring in your resident dog to
-            meet the prospective new family member. This is to make sure there
-            will be no future dog-to-dog conflicts.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            The best municipal agencies will be very like a private shelter.
-            Others will simply give you access to their kennel areas and let you
-            make your own choice. They may not have a special “get-acquainted”
-            area, but you should be able to interact with the pet enough to know
-            if you’ve met your match.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            What kind of paperwork will I have to fill out at a private or
-            municipal shelter?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            All private shelters will ask you to fill out an adoption
-            application.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Sometimes they do this before you’ve met any pets, and other times,
-            they don’t have you do it until you’ve selected a pet you’d like to
-            adopt. Sometimes, the questionnaire is mostly about practical things
-            like your experience with pets, landlord information, and
-            references. Other times, it’s part of the process by which the
-            shelter staff help steer you toward the pet who is right for your
-            lifestyle.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Municipal shelters sometimes have the same paperwork requirements as
-            private shelters, but usually it’s far more streamlined.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            Should I choose a private or municipal shelter? Is one better than
-            the other?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            However un-fancy the facility or minimal the staff, pets in animal
-            control agencies and municipal shelters are just as wonderful as
-            pets in other types of organizations. The adoption experience might
-            be more pleasant at a private shelter, and the adoption fee is
-            frequently much lower at a municipal shelter, but in the end, it’s
-            about finding your special pet, wherever they are!
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            The pet I am interested in is with a rescue group, and I’ve emailed
-            them about adopting him. What will happen next?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Because most rescue groups are run by volunteers, it can take a
-            little longer for them to get back to you about adoption than a
-            “brick and mortar” shelter or animal control agency.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Rescue groups’ shelter pets are usually fostered by organization
-            members. After inquiring, you’ll often be invited to meet both the
-            foster parent and shelter pet at a central adoption area, such as an
-            event at a pet supply store.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            What are the pros and cons of adopting from a rescue group?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Rescue group adoptions tend to be much more personalized. The foster
-            families know the pets very well – they should, since they live with
-            them! – and they can give you a lot more information about how the
-            pet fits into a family or home than shelters usually can.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            This personalization often makes the adoption process more rigorous,
-            and it can take a lot longer. The rescue group may want to conduct a
-            home visit before finalizing the adoption, and will probably contact
-            your veterinarian and references, as well.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            Will I be able to take my pet home immediately?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Usually, yes. Sometimes you’ll have to wait for the pet to be spayed
-            or neutered.
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            How much will it cost?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            While some pets are free, adoption fees commonly range between $25
-            and $250. If you feel the cost is too high, ask what it’s based on.
-            Was the pet with a rescue group that receives no outside funding,
-            pays for all veterinary care, spay/neuter, training, housing, etc.
-            out of their own pockets? Did the pet have extraordinary expenses,
-            such as costly medical care, before being made available for
-            adoption?
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            What happens if there are problems when I get my pet home?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Some shelters and rescue groups have free or discounted services,
-            such as dog-training classes or veterinary or behavior assistance,
-            to make sure your adoption goes smoothly after you get your new pet
-            home.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Rescue groups usually provide a lot of follow-up to adoptions, and
-            some shelters do as well. If you’re having health, behavior or
-            adjustment problems with your new pet, call your adoption location
-            and ask if they have any services that can help things work out!
-          </p>
-          <h2 class="font-inter font-semibold text-base text-brown-900 px-7 mb-4">
-            After my pet has settled in and we’re loving each other – what do I
-            do next?
-          </h2>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            someone else who has. So your next step is to tell the world about
-            your wonderful pet, and where you found him or her!
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-32">
-            And, of course, to live happily ever after!
-          </p>
-        </div>
-        <hr />
-      </div>
-      <div>
-        <FooterSection />
-      </div>
-    </div>
-  );
-};
-
-export default FAQsPage;
