--- src/pages/links/termsAndConditionsPage.js
@@ -1,719 +1,0 @@
-import React from "react";
-import FooterSection from "../components/FooterSection";
-
-const TermAndConditions = () => {
-  return (
-    <div class="mx-auto bg-dark-light">
-       <div class="flex justify-center max-w-[100px] sm:max-w-[1120px]">
-        <img
-          src="./images/logo_r_brown.svg"
-          alt="logo_brown"
-          className="logo-r-brown"
-        />
-      </div>
-      <hr />
-      <div class="flex justify-center">
-        <div class="max-w-4xl">
-          <h1 class="font-inter font-bold text-4xl text-brown-900 p-7">
-            Swipet Terms of Service
-          </h1>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            This page contains our site policies for the following:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            GENERAL TERMS AND CONDITIONS
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            COPYRIGHT INFRINGEMENT NOTIFICATION
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            USER GENERATED CONTENT TERMS AND CONDITIONS
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            LINKING POLICY
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            SHELTER MEMBER- ADDITIONAL TERMS AND CODE OF CONDUCT
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Thank you for visiting this website (the “Site” or “website”).
-            Please read the Terms and Conditions contained in this document
-            carefully since any use of this site constitutes your acceptance of
-            the Terms and Conditions set out herein. If you do not agree to
-            these terms and conditions, please do not use this website.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Throughout this site, the terms “we”, “us”, “our”, and “Petfinder”
-            refer to Petfinder, a division of Nestlé Purina PetCare Company, a
-            Missouri corporation. “You” refers to any person accessing and/or
-            using this website.
-          </p>
-
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Please Note: These terms and conditions may be revised by Petfinder
-            by updating this posting – Please read them carefully.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            1. Privacy Notice
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Any personal information or material sent to this website are
-            subject to the terms and conditions set out in our Privacy Policy.
-            We encourage you to read and refer to it before you submit any
-            personal information to this Site. The Privacy Policy is a part of
-            these terms and conditions.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4"></p>
-
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            2. Accuracy, completeness and timeliness of information
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            We are not responsible if the information that we make available on
-            this website is not accurate or complete. Any reliance upon the
-            material on this website shall be at your own risk. You agree that
-            it is your responsibility to monitor any changes to the material and
-            the information contained on this website.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            3. Submissions to this Site
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Any non-personal communication or material you submit to this
-            website by electronic mail, upload, or otherwise, including any
-            data, questions, comments, suggestions or the like (“Submissions”)
-            are, and will be treated as, non-confidential and non-proprietary.
-            Anything you submit or post becomes the property of Petfinder and
-            Petfinder may use and reproduce such Submission freely, and for any
-            purpose. Specifically, Petfinder is free to use for any purposes
-            whatsoever (including but not limited to the developing,
-            manufacturing, advertising and marketing of products) any ideas,
-            artwork, inventions, developments, suggestions or concepts contained
-            in any Submission you send to this website. Such use shall not
-            include advertising targeted to breeders. Any such use is without
-            compensation to the party submitting information. By making a
-            Submission, you are also warranting that you own the
-            material/content submitted, that it is not defamatory and that
-            Petfinder’s use will not violate any third party’s rights or
-            otherwise place us in breach of any applicable laws. Petfinder is
-            under no obligation to use the information submitted.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            4. Intellectual Property Rights
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            All copyright, trademarks and other intellectual property rights in
-            all text, images and other materials on this website are the
-            property of Petfinder or are included with the permission of the
-            relevant owner.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You are permitted to browse this website, reproduce extracts by way
-            of printing, downloading to a hard disk or for the purposes of
-            distribution to other individuals. This is only to be done on the
-            proviso that you keep intact all copyright and other proprietary
-            notices and that the trademark notice below appears on such
-            reproductions. No reproduction of any part of this website may be
-            sold or distributed for commercial gain nor shall it be modified or
-            incorporated in any other work, publication or website.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            ® Reg. Trademark of Société des Produits Nestlé S.A. All rights
-            reserved.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            The trademarks, logos, characters and service marks (collectively
-            “Trademarks”) displayed on this website belong to Société des
-            Produits Nestlé S.A which licenses such use to Petfinder. Nothing
-            contained on this website should be construed as granting any
-            license or right to use any Trademark displayed on this website.
-            Your use/misuse of the Trademarks displayed on this website, or on
-            any other content on this website, except as provided for in these
-            Terms and Conditions, is strictly prohibited. You are also advised
-            that Petfinder will vigorously enforce its intellectual property
-            rights to the fullest extent of the law.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            5. Links to Other Websites
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Links on the Petfinder websites may take you outside the Petfinder
-            network and systems and Petfinder accepts no responsibility for the
-            content, accuracy or function of these other third party websites.
-            The links are provided in good faith and Petfinder cannot be held
-            responsible for any subsequent change in other third party websites
-            to which we provide a link. The inclusion of any link to other
-            websites does not imply endorsement by Petfinder. We highly
-            recommend that you make yourself aware of and carefully read the
-            legal and privacy notices of all other websites that you visit.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            6. Warranties and Disclaimers
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Your use of this website is at your own and exclusive risk.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Warranties
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            THIS WEBSITE IS PROVIDED TO YOU ON AN “AS IS” AND “AS AVAILABLE”
-            BASIS AND, CONSEQUENTLY, PETFINDER MAKES NO WARRANTIES OF ANY KIND,
-            WHETHER EXPRESS, IMPLIED, STATUTORY OR OTHERWISE, INCLUDING BUT NOT
-            LIMITED TO THE IMPLIED WARRANTIES OF MERCHANTABILITY OR SATISFACTORY
-            QUALITY AND FITNESS FOR A PARTICULAR PURPOSE) INCLUDING WARRANTIES
-            OR REPRESENTATIONS THAT MATERIAL ON THIS WEBSITE WILL BE COMPLETE,
-            ACCURATE, RELIABLE, TIMELY, NON-INFRINGING TO THIRD PARTIES, THAT
-            ACCESS TO THIS WEBSITE WILL BE UN-INTERRUPTED OR ERROR-FREE OR FREE
-            FROM VIRUSES, THAT THIS WEBSITE WILL BE SECURE, THAT ANY ADVICE OR
-            OPINION OBTAINED FROM PETFINDER THROUGH THIS WEBSITE IS ACCURATE OR
-            TO BE RELIED UPON AND ANY REPRESENTATIONS OR WARRANTIES THERETO ARE
-            ACCORDINGLY EXPRESSLY DISCLAIMED.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Please note that some jurisdictions may not allow exclusions of
-            implied warranties, so some of these exclusions may not apply to
-            you. Please check your local laws.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            We reserve the right to restrict, suspend or terminate without
-            notice your access to this website or any feature of this website or
-            any feature or part thereof at any time.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            7. Pet Health, Fitness and Nutrition
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            This Site may from time to time contain information relating to
-            nutrition and various medical, health and fitness conditions of pets
-            and their treatment. This is for informational purposes only and is
-            not meant to be a substitute for the advice provided by your own
-            veterinarian or other veterinary professional. You should not use
-            the information contained herein for diagnosing a pet’s health or
-            fitness problem or disease. You should always consult your own
-            veterinarian and/or veterinary advisors.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            8. Limitation of Liability
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder and/or any other party involved in the creating, producing
-            or delivering this website on our behalf shall have no liability or
-            any responsibility whatsoever for any direct, incidental,
-            consequential, indirect, special or punitive damages, costs, losses
-            or liabilities whatsoever and howsoever arising out of your access
-            to, use, inability to use, change in content of this website or
-            arising from any other website you access through a link from this
-            website or to the extent permitted by applicable law, from any
-            actions we take or fail to take as a result of any electronic mail
-            messages you send us.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder shall have no responsibility to maintain the material and
-            services made available on this website or to supply any
-            corrections, updates, or releases in connection therewith. Any
-            material on this website is subject to change without notice.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Further, Petfinder shall have no liability or any responsibility
-            whatsoever for any loss suffered caused by viruses that may infect
-            your computer equipment or other property by reason of your use of,
-            access to or downloading of any material from this website. If you
-            choose to download material from this website you do so at your own
-            risk.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Although Petfinder may from time to time monitor or review
-            discussions, chats, postings, transmissions, bulletin boards, and
-            the like on the Site, Petfinder is under no obligation to do so and
-            assumes no responsibility or liability whatsoever arising from the
-            content of any such locations including but not limited to any
-            error, defamation, libel, slander, omission, falsehood, obscenity,
-            pornography, profanity, danger or inaccuracy contained in any
-            information within such locations on the Site.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            To the maximum extent permitted by applicable law, you expressly
-            waive all claims against Petfinder, its officers, directors,
-            employees, suppliers and programmers that may arise from your use or
-            access of this website.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            9. Prohibited Activities
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You are prohibited from doing any act that Petfinder in its absolute
-            discretion may deem to be inappropriate and/or would be deemed to be
-            an unlawful act or is prohibited by any laws applicable to this
-            website including but not limited to:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            • Any act that would constitute a breach of either the privacy
-            (including uploading private information without the concerned
-            individual’s consent) or any other of the legal rights of
-            individuals
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            • Using this website to defame or libel Petfinder, its employees or
-            other individuals or acting in such a way that brings into disrepute
-            the good name of Petfinder;
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            • Uploading files that contain viruses that may cause damage to the
-            property of Petfinder or the property of other individuals; and
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Posting or transmitting to this website any non-authorized material
-            including but not limited to material that is in our opinion likely
-            to cause annoyance, detrimental to or in violation of Petfinder’s or
-            any third party´s systems or network security, libelous, defamatory,
-            racist, obscene, threatening, pornographic, profane or is otherwise
-            unlawful.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            10. Export Restrictions
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Software from this Site may be further subject to United States
-            Export laws, rules and regulations, as amended from time to time. No
-            software from this Site may be downloaded or exported (i) into (or
-            to a national or resident of) Cuba, Iraq, Libya, North Korea, Iran,
-            Syria, Sudan or any other country to which the United States has
-            embargoed goods; or (ii) anyone on the United States Treasury
-            Department’s list of Specially Designated Nationals or the U.S.
-            Commerce Department’s Table of Deny Orders. By downloading or using
-            the software, you represent and warrant that you are not located in,
-            under the control of, or a national or resident of any such country
-            or on any such list.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            11. Mobile Services
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            If Petfinder permits or makes available to you the ability to: (a)
-            upload content to the website via your mobile telephone and/or
-            tablet (‘mobile device”), (b) receive and reply to messages, or to
-            access or make posts using text messaging, (c) browse the site from
-            your mobile device or (d) to access features through a mobile
-            application you have downloaded and installed on your mobile device
-            (collectively “Mobile Services”), your mobile carrier’s standard
-            data, Internet, messaging and other service fees and charges apply
-            to your use of, and access to, the Mobile Services, and you are
-            solely responsible for all such fees and charges. The Mobile
-            Services may not be compatible with all mobile devices, carriers,
-            and services, and Petfinder does not make any representations or
-            warranties regarding such compatibility. You are solely responsible
-            for ensuring that your mobile device is compatible with the type of
-            content being provided through the Mobile Services. You understand
-            that content delivered as part of the Mobile Services may be
-            delivered to your device through premium mobile services and
-            acknowledge that you are solely and exclusively responsible for any
-            charges for such services, including but not limited to any roaming
-            charges.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You may be given the option to register to receive commercial
-            messages from Petfinder via your mobile device. You understand that
-            you may be charged by your wireless service provider in connection
-            with the receipt of such messages. You may “opt out” of receiving
-            future messages by following the instructions provided (i) in the
-            wireless messages or (ii) on the website where you registered to
-            receive such messages.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            12. Jurisdiction and governing law
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            This website is located in and operated from the United States of
-            America. Your use and access to this website will be governed by the
-            laws of the United States. For information regarding how we treat
-            data collection and the transfer of personally identifiable
-            information from countries or jurisdictions outside the US, please
-            review our Privacy Policy.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You and Petfinder agree that any controversy or claim arising from
-            or pertaining to the use of this website shall be governed by the
-            laws of the State of Missouri, United States of America, without
-            giving effect to its choice-of-law provisions, and be submitted to
-            the exclusive jurisdiction of the courts of St. Louis County, State
-            of Missouri, United States of America.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            13. Legal notice update
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            We reserve the right to make any changes and corrections to this
-            notice. Please refer to this page from time to time to review these
-            and new additional information.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            ________________________________________
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            COPYRIGHT INFRINGEMENT NOTIFICATION
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder is committed to complying with U.S. copyright law and to
-            respond to claims of copyright infringement. Petfinder will promptly
-            process and investigate notices of alleged infringement and will
-            take appropriate actions under the Digital Millennium Copyright Act,
-            Title 17, United States Code, Section 512(c) (“DMCA”).
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Pursuant to the DMCA, notifications of claimed copyright
-            infringement should be sent to a Service Provider’s Designated
-            Agent. Notification must be submitted to the following Designated
-            Agent for this site in the manner described below:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            By Mail:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            Nestlé Purina PetCare
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            ATTN: Legal Department, DMCA Copyright Notices
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            Checkerboard Square
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            St. Louis, MO 63164
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7">
-            By Email:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            copyrightnotice@petfinder.com
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            For your complaint to be valid under the DMCA, you must provide all
-            of the following information when providing notice of the claimed
-            copyright infringement:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            1. A physical or electronic signature of a person authorized to act
-            on behalf of the copyright owner;
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            2. Identification of the copyrighted work claimed to have been
-            infringed;
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            3. Identification of the material that is claimed to be infringing
-            or to be the subject of the infringing activity, and that is to be
-            removed or access to which is to be disabled, as well as information
-            reasonably sufficient to permit Petfinder to locate the material;
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            4. Information reasonably sufficient to permit Petfinder to contact
-            the copyright owner, such as an address, telephone number, and, if
-            available, an electronic mail address;
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            5. A statement that you have a good faith belief that use of the
-            material in the manner complained of is not authorized by the
-            copyright owner, its agent, or law; and
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            6. A statement that the information in the notification is accurate,
-            and under penalty of perjury, that you are authorized to act on
-            behalf of the copyright owner.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            For more details on the information required for valid notification,
-            see 17 U.S.C. 512(c)(3).
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You should be aware that, under the DMCA, claimants who make
-            misrepresentations concerning copyright infringement may be liable
-            for damages incurred as a result of the removal or blocking of the
-            material, court costs, and attorneys’ fees.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            ________________________________________
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            USER GENERATED CONTENT TERMS AND CONDITIONS
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder owns and/or operates various websites that permit users
-            (hereinafter “Members” or “Users”) to submit or publish content
-            (“User-Generated Websites”). Such websites may include, but are not
-            limited to, message boards, blogs, social networks, and comment
-            boards. By registering or using any User-Generated Websites, you
-            accept and agree to the following Terms and conditions for
-            User-Generated Content.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Eligibility and registration
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            As part of the registration process, you will be asked to select a
-            username and password and you will be responsible for all activities
-            occurring under your username and for keeping your password secure.
-            We may refuse to grant you a username that impersonates someone
-            else, is or may be protected by trademark or proprietary rights law,
-            or is vulgar, offensive or otherwise inappropriate, as determined by
-            us in our sole discretion. You are solely responsible for any and
-            all uses of your username, password, and account.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Appropriate conduct and content; personal use only User-Generated
-            Websites are made available for your personal, non-commercial use
-            only. User-Generated Websites may not be used for any commercial or
-            promotional purpose.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            By acceptance of these User-Generated Websites Terms and Conditions,
-            you agree that any information you provide about yourself and your
-            pets is accurate and complete information at the time provided. In
-            particular, you represent that you are 13 years of age or older. In
-            addition, you agree to update any information provided to
-            User-Generated Websites to keep it accurate and complete. If you
-            become a member of a User-Generated Website, your failure to
-            accurately and completely provide, and timely update, information
-            about yourself is reasonable grounds for suspension or termination
-            of your membership.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You may not use User-Generated Websites for any illegal or
-            unauthorized purpose. International users agree to comply with all
-            local rules regarding online conduct and acceptable content,
-            including laws regulating the export of data from the United States
-            or your country of residence. You are solely responsible for your
-            conduct and any data, text, information, photos, links and other
-            content (“Materials”) that you submit, post, and display on
-            User-Generated Websites.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            We may without notice to you, but shall have no obligation to,
-            refuse, edit, and/or remove Materials that we determine in our sole
-            discretion are unlawful, fraudulent, threatening, libelous,
-            defamatory, obscene or otherwise objectionable, or infringing or
-            violating any party’s intellectual property or other proprietary
-            rights or these User-Generated Websites Terms and Conditions. We
-            also reserve the right in our sole discretion to refuse, edit,
-            and/or remove any content that may be deemed by any User-Generated
-            Website community to be inappropriate without providing you notice.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Children’s privacy
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            We are committed to protecting the privacy of children. The
-            User-Generated Websites are is not intended or designed to attract
-            children under the age of 13. We do not collect personally
-            identifiable information from any person we have actual knowledge is
-            under the age of 13.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Examples of illegal or unauthorized uses include, but are not
-            limited to:
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Modifying, adapting, translating, or reverse engineering any portion
-            of the User-Generated Websites; using any robot, spider, site
-            search/retrieval application, or other device to retrieve or index
-            any portion of User-Generated Websites; collecting any information
-            about other Members (including usernames and/or e-mail addresses)
-            for unauthorized purposes; reformatting or framing any portion of
-            the web pages or site images that are part of User-Generated
-            Websites; creating user accounts by automated means or under false
-            or fraudulent pretenses; creating or transmitting unwanted
-            electronic communications such as “spam,” or chain letters to other
-            Members or otherwise interfering with other Member’s enjoyment of
-            the service; submitting Materials of any third party without such
-            third party’s prior written consent; submitting Materials that
-            falsely express or imply that such Materials are sponsored or
-            endorsed by User-Generated Websites; submitting Materials that
-            infringe, misappropriate or violate the intellectual property,
-            publicity, privacy or other proprietary rights of any party;
-            transmitting any viruses, worms, defects, Trojan horses or other
-            items of a destructive nature; submitting Materials that are
-            unlawful or promote or encourage illegal activity; displaying an
-            advertisement as part of your or your pet’s profile; any automated
-            use of the system, including the use of scripts to send messages or
-            post comments; or submitting false or misleading information.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            While Petfinder prohibits such conduct and content on its
-            User-Generated Websites, you understand and agree that you
-            nonetheless may be exposed to such Materials and that you use
-            User-Generated Websites at your own risk.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Changes to Terms and Conditions for User-Generated Content
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            We reserve the right to modify or terminate any User-Generated
-            Website or service therein for any reason, and without notice,
-            without liability to you, any other Member, or any third party. We
-            also reserve the right to modify these Terms and Conditions for
-            User-Generated Content from time to time without notice. You are
-            responsible for regularly reviewing these Terms and Conditions for
-            User-Generated Content so that you will be apprised of any changes.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            License to Use Materials
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            By submitting, posting, or displaying any Materials on or through
-            User-Generated Websites, you automatically grant Petfinder a
-            worldwide, non-exclusive, royalty-free, irrevocable license to
-            publicly display or re-display such Materials on the User-Generated
-            Websites website and in all other forms of media (including, without
-            limitation, e-mail blasts and mobile or text messages or
-            applications) that are used by Petfinder or its agents in connection
-            with the promotion, marketing, use, or expansion of any
-            User-Generated Websites.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            You represent to Petfinder that you own all Materials you provide
-            and/or have the unencumbered legal right to grant the
-            above-referenced license with respect to all Materials you provide.
-            You further represent to Petfinder that such materials do not
-            infringe any third party copyrights or other third party
-            intellectual property rights. You agree to indemnify Petfinder for
-            any and all loses resulting from the invalidity or inaccuracy of
-            such representation.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Except as expressly authorized by User-Generated Websites, you may
-            not copy, modify, publish, transmit, distribute, perform, display or
-            sell any of Petfinder’s proprietary information, which includes the
-            sum of all postings and photos to the site.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Please do not send to any User-Generated Website or post on any
-            User-Generated Website forum anything that constitutes a trade
-            secret, confidential or proprietary information. We are not in a
-            position to accept such information, nor can we agree to obligations
-            of nondisclosure or confidentiality with regard to submitted plans
-            or ideas. You agree that any information or materials that you, or
-            individuals acting on your behalf, provide to any User-Generated
-            Website will not be considered confidential or proprietary.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            User-Generated Websites includes information and content generated
-            from a wide variety of sources, including board certified
-            veterinarians, trainers, and other professional service providers,
-            as well as, other Users. Users should understand that User-Generated
-            Websites is a place to review and share information — but the
-            information has not been verified for accuracy and does not
-            represent the opinions or positions of Petfinder. Petfinder cannot
-            guarantee the timeliness, accuracy, efficacy, or veracity of any
-            information provided at User-Generated Websites, nor the efficiency
-            or applicability of any advice, opinions or content as they may be
-            interpreted and applied to your pet. Petfinder’s personnel are not
-            and will not be responsible for any misleading, false, or otherwise
-            injurious information and advice communicated on User-Generated
-            Websites, or for any results obtained from the use of such
-            information or advice. Petfinder and its personnel will not be
-            liable for any loss or damage suffered by a User through the User’s
-            reliance on information, content and advice appearing on our
-            website.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            USER-GENERATED WEBSITES ARE BEING MADE AVAILABLE TO YOU “AS IS,”
-            WITHOUT REPRESENTATIONS OR WARRANTIES OF ANY KIND.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Standard Terms and Conditions for Petfinder Sites
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            In addition to these User-Generated Websites Terms and Conditions,
-            you agree to be bound by Petfinder’s standard Terms and Conditions
-            for its websites, as the same may be modified and updated from time
-            to time, with respect to your use of User-Generated Websites.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            ________________________________________
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            LINKING POLICY
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder grants a nonexclusive, limited permission (“Permission”)
-            to display the Trademarks contained in Petfinder domain names, brand
-            names and logos to the owner(s) of any web sites (“Owner”) that
-            would like to develop a link between their site and the
-            corresponding web site owned by Petfinder (“Petfinder Sites”). The
-            materials on Owner’s site shall not imply any association with or
-            relation to Petfinder or imply in any way that any materials from
-            Petfinder Sites are maintained within Owner’s web site.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            No changes or modifications to the Petfinder domain names, brand
-            names or logos may be made. Use of the Petfinder domain names, brand
-            names or logos for purposes other than linking to Petfinder Sites is
-            expressly forbidden. The domain names, brand names and logos must
-            not be used or associated with any obscene, scandalous,
-            inflammatory, pornographic, indecent, profane, defamatory or
-            unlawful site or materials in any way, or in any manner that would
-            adversely affect the name, reputation and goodwill of Petfinder and
-            its products. Petfinder retains and reserves all rights in and to
-            Petfinder ‘s trade names, trademarks, copyrights and all other
-            intellectual property with respect to its domain names, brand names
-            and logos. Petfinder reserves the right to cancel the Permission at
-            any time and for any reason. Upon receipt of such notification, use
-            of the domain names, brand names or logos must cease and the link
-            between the web sites must terminate. Use of Petfinder trademarks,
-            trade names, logos, copyrighted material or any other intellectual
-            property of Petfinder for any purpose not stated herein is expressly
-            forbidden.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            All materials are provided “”AS IS,”” WITHOUT WARRANTY OF ANY KIND,
-            EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
-            WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, OR
-            NONINFRINGEMENT. Please note that some jurisdictions may not allow
-            the exclusion of implied warranties, so some of the foregoing
-            exclusions may not apply to you. Check your local laws for any
-            restrictions or limitations regarding the exclusion of implied
-            warranties.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder shall not be responsible for any loss, damage, liability
-            or expense that may accompany or result from your use of links,
-            domain names, brand names or logos, including but not limited to any
-            computer virus; technical, hardware or software failures of any
-            kind; lost or unavailable network connections; or failed,
-            incomplete, garbled or delayed computer transmissions. The granting
-            of this Permission does not indicate nor should it be construed in
-            any way to represent an endorsement by Petfinder of any materials on
-            a web site containing a Petfinder domain name, brand name or logo or
-            any other Petfinder intellectual property.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder controls and operates its web sites from the company’s
-            North American pet products headquarters in St. Louis, Missouri, in
-            the United States of America. Petfinder in no way implies that the
-            domain names, brand names or logos are appropriate or available for
-            use outside of the United States. If you use our materials from
-            locations outside of the United States, you are responsible for
-            compliance with any applicable laws.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-4">
-            Petfinder may at any time revise this Linking Policy. You are bound
-            by any such revisions and should therefore periodically visit this
-            page to review the then-current Linking Policy to which you are
-            bound.
-          </p>
-          <p class="font-inter  font-normal text-base text-brown-900 px-7 mb-32">
-            The display of Petfinder domain names, brand names and logos on the
-            web site indicates the agreement by the web site Owner to the
-            foregoing terms and conditions.
-          </p>
-        </div>
-      </div>
-      <hr />
-      <div>
-        <FooterSection />
-      </div>
-    </div>
-  );
-};
-
-export default TermAndConditions;
